Fix suppress* tests on Vista+ when ntdll.pdb is not present.
Fixes issue 577.

The test output doesn't match on my system or the buildbot because the
block is still reachable from RtlpHeapFailureInfo, which was covered by
i#292.  However, the buildbot doesn't have ntdll.pdb, and until we find
a way to pull it in as part of the test suite, we'll accept either
output.  The goal of this test is not to verify that we report a leak
after attempting to free the middle of an object, but to test our
detection of invalid heap arguments to free.

SVN-Revision: 573