Fix the nudge test on Windows by accepting ~~(Dr\.M|[0-9]+)~~ for the console prefix.
Fixes issue 653.

We also make the test more strict.  On all platforms, we should see at
least 2 error summaries from the two nudges.  On Linux, we get a third
because the frontend scrapes the log files to put together the summary.
On Windows, we will the application with no chance for recovery, so we
don't get an at exit summary.

SVN-Revision: 578