Dump stdout/stderr in runtest.cmake when matching fails.
This allows us to get more information from logs of failing tests,
because the raw stdout/stderr is not recorded anywhere else.

I also fixed the matching code that was intended to remove up to and
including the matching line to remove the line that matched.  We rely on
this behavior in the nudge test on Linux to count the number of error
reports.

SVN-Revision: 594