Handle the new networking ioctls used by IPHLPAPI.dll!GetAdaptersInfo.
I was unable to find any ioctl codes that matched the values seen
here, so I had to reverse engineer the in/out structure used for these
two ioctls.

The 0x12000f / 0x003 ioctl has one input parameter and two output parameters,
and they are always on the stack.

The 0x12001b / 0x006 ioctl has up to 4 output parameters, all of which
are allocated on the heap when it is called through NSI.dll.  This was
convenient, because I was able to easily confirm that the sizes of
these buffers always match an element size field times an element
count field.  The vast majority of the uninit false positives came
from marking these buffers as being written.

Fixes issue 719.

SVN-Revision: 695