i#838: Avoid taking symbolized stack traces in a suppressed module.
Before taking the callstack, we first lookup the module associated with
the current PC, and run through the suppression list for the error type,
looking for a single-frame suppression covering a matching module.

This could be optimized to avoid the extra suppression list iteration by
hooking the module load event and storing a suppress_spec_t pointer in
the module area rb tree in callstack.c, or maintaining an independent rb
tree.

SVN-Revision: 809