i#511: Mark the output of UuidCreate as defined
Fixes issue 511.

UuidCreate uses the uninitialized bits of its output buffer to seed its
random number generation, and we propagate the definedness through to
the output.  This causes uninits when the app tries to read the
generated GUID.

This CL introduces a wrapper for UuidCreate which marks the GUID as
defined if the call succeeds.

Review URL: https://codereview.appspot.com/6492056

SVN-Revision: 1004