i#946 part 2: socketcall for x64
It's not easy to share all the code between SYS_socketcall and x64's nicer
top-level split.  I went with special arg type handling for sockaddr and
msghdr.  For sockaddr I call the existing check_sockaddr() (augmented to do
the kernel length adjustment in-house) and for msghdr I pulled the code out
of the socketcall sequence in order to share it.

Added a sorely-needed socket test to app_suite.  Confirmed manually
that it finds real errors in the receive buffer.

Added an extra_info array for use by syscall parameters that need to store
information from pre to post.  Folded the sysarg_size_from_field data into
the extra_info array.

SVN-Revision: 1028