i#1040: add non-ASCII support to Dr. Memory front-end
To finish off i#1040 we need to finish DRi#943.
This issue gets Dr. Memory's front-end working, and with the DR core
portion of DRi#943, it fixes the particular bug raised in the issue.

For unicode support, we must use wide-char versions of main(), argv[],
and all of the windows library routines like GetEnvironmentVariable(),
_access(), and GetFullPathName().  Yet we must use UTF-8 for the DR
API routines.  We could go two routes: one is to have everything be UTF-16
and only convert before calling DR routines; the other is to have everything
be UTF-8 and convert when calling Windows routines.  We pick the latter
because we expect to port this to Linux.

I tried to add a test that uses a unicode-named logdir but I can't get the
non-ascii chars through execute_process() on Windows.

SVN-Revision: 1111