fixes issue 1119
For general syscall handling, if the kernel-written size of a param is
larger than the passed-in size, use the passed-in size.

Changed check_sockaddr() to check the whole struct for addressability,
which was missing for many 32-bit cases.  For a write, the kernel seems to
fill in the size it specified w/o gaps, so not doing a per-field walk in
that case.

Added a getsockname test to drsyscall_app.c and a check in
drsyscall_client.c.  The strace test output also shows the fix but doesn't
programmatically verify it.

SVN-Revision: 1161