fixes issue 1148
Breaks drsyscall compatibility: but drsyscall has not been officially
released yet so that's allowed.  This commit message is a heads up to
anyone using it now.

Adds is_byte_undefined() to drsys_options_t, because undefined != !defined
and we want to avoid the perf hit of two shadow lookups.

Also relaxes assert that could fire in a race to a warning.

SVN-Revision: 1209