i#772: win8 support: system calls
Adds win8 win32k.sys system call numbers, which need to be split into 3 and
not the normal 2: x64, wow64, and x86.
Adds the new syscalls to the win32k.sys tables.
Adds the new ntoskrnl syscalls to that table and fills in two known ones.
Adds new arrays and selects them on win8.

Usercall numbers are separated into another commit.

i#1153 covers filling in the rest of the details.

SVN-Revision: 1215