i#772: windows 8 support: NtUserBuildHwndList
Added a new feature of Windows-version-specific syscall table entries, by
using the number field as a min and a max version.

Used the feature to handle NtUserBuildHwndList which has a new arg inserted
in the middle on Win8.

Also marked 2 syscalls removed in win8

Did not yet mark all the minimums for syscalls added in certain versions as
it's not all that useful.

SVN-Revision: 1217