fixes issue 1165
The post-call wrapping symcache entries end up occupying a long linked
list which is traversed both on lookup iteration and on every add
while parsing the symcache file.  This shows up in profiling of unit_tests
and browser_tests.  The solution is to use a hashtable and a linked
list and to cache the last iter entry.

SVN-Revision: 1241