i#1176 part 1: Linux multi-arena support
Properly supports multiple arenas once the brk fills up.
Uses a simple design of always using the top-level pointer and walking all
the sub-arenas every time as the overhead does not appear to be enough to
warrant a more complex solution.
This also fixes latent issues with Windows support for multiple regions for
one Heap.

SVN-Revision: 1258