i#948: add coalescing, step 3: coalesce + return memory
Implements forward and backward coalescing.

Reverses the brk and de-allocates arena regions if coalescing results in a
large enough memory region.

For large coalesced chunks I had to store the prev size outside of the
header, which impacts pattern mode but it seems worthwhile.

Added a malloc stress test of coalescing and splitting.

SVN-Revision: 1262