fixes issue 1184 fixes issue 779
Adds pattern-mode support to -replace_malloc.
Also fixes i#779: pattern support for -no_replace_realloc.

Instead of using a hashtable and no co-located header (that's i#879), or
using a single co-located pointer, decided to implement the simplest thing:
don't share redzones, put the header between the redzones, and just don't
have the pattern value on the header.  This matches the current wrap
behavior, in any case, and is the smallest change to the current
-replace_malloc allocator.

Rather than storing the free list next pointer in the redzone, and
complicating pattern mode's redzone writes, the allocator extends the live
header an extra 8 bytes.

Augmented the alloc.h interface:
+ Added more args to client_handle_realloc()
+ Added client_handle_free_reuse() to properly handle the -replace_malloc
  allocator doing its own free delay.
+ Added client_new_redzone() to handle the -replace_malloc allocator
  splitting and coalescing free chunks, removing and creating redzones
  between free chunks.
+ Added for_reuse to client_handle_free() to handle the -replace_malloc
  allocator's greater control over delayed or immediate-reuse frees.

Refactored some of the pattern code.

SVN-Revision: 1279