Fixes issue 1232
Relax assert that assumed padding was smaller than chunk alignment, which
is not true for a realloc-shrunk object.

Add reporting of an unaddr access that is in the redzone or padding of a
freed object as "%d byte(s) beyond memory that was freed".

Also removed the -1 adjustment on the byte offset reported for adjacent
mallocs, as it's more confusing than it helps ("0 bytes beyond" seems
reasonable).

Added a test of the new reporting to free.c.

SVN-Revision: 1386