i#1247 part 1: NtPowerInformation.PowerRequestCreate
Add special handling for NtPowerInformation.PowerRequestCreate, which
has an unknown struct (added to wdm.h) whose input fields are defined
by its Flags field.  Includes handling of the unknown flag 0x80000000
which seems to require no other fields to be initialized.

Adds a new test of PowerCreateRequest, which invokes NtPowerInformation.

Not everything is understood here: some data from REASON_CONTEXT seems
to be lost on the path to the kernel.

SVN-Revision: 1391