i#1265 part 1: fix callstack issues
For callstacks at a syscall on 32-bit Linux on a 64-bit kernel, skip the
vsyscall sysenter "push ebp" to avoid skipping over a frame during the
callstack walk, as the libc routine that invoked the syscall often doesn't
have a frame pointer.

Adds drsys_syscall_gateway() to the Dr. Syscall API.

SVN-Revision: 1419