i#1306 Dr. Heapstat -replace_malloc, part 1:
Pass in the client library name via preprocessor define CLIENT_LIBNAME.
Generalize the existing name check in the alloc module load event to
skip the drheapstat library as well as the drmemory library.

Fixes mistakes in setting properties of the wrong lib in dstracelib, which
were blocking the change here.

SVN-Revision: 1481