Fixes issue 1310
Adds new options -report_whitelist and -whitelist_num_frames.  Only if one
of the top -whitelist_num_frames frames of an error callstack matches one
of the -report_whitelist patterns in its module path is an error reported
normally; else it is separated as a potential error just like with the
blacklist.

Adds two new tests for the whitelist.

Updates the documentation with a new section on the blacklist and
whitelist.

Changes the default blacklist to $SYSTEMROOT*.d?? to match .drv and .dll.

The -pause_at_* options do not stop at potential errors.

SVN-Revision: 1512