Fixes issue 1323
Adds a warning on a non-zero exit, for Windows only and printed from
frontend.c.

To make it look nice in VS I ended up having to parse -prefix_style in
frontend which is a little ugly but I like the prefix appearing on this
particular message, just like it does on the symbol loading messages.

SVN-Revision: 1527