i#1346 Windows 8.1 support: update usercalls
Windows 8.1 does in fact add quite a few usercalls.
i#1360 covers identifying the added ones: for now all we can do is shift
the numbers to match the points of calls we see in our test apps (primarily
app_suite_tests and calc).

Fixes an incorrect number sequence for Windows 8 as well.

SVN-Revision: 1602