Fixes issue 1377
Add a check for supported Windows versions to avoid surprises and
mysterious crashes on future releases.

Check the return value of dr_standalone_init() and consider NULL to
mean we're on an OS unsupported by DR.  Requires DRi#1317 == r2371.

Switch to using dr_get_os_version(), as we can't trust GetVersionEx()
(it pretends Win8.1 is Win8 and is now deprecated).

SVN-Revision: 1619