i#703 callstack quality improvements:
Add two new options, -no_callstack_use_fp and -callstack_conservative, for
improving callstack quality with mixed FPO and unoptimized code.

-no_callstack_use_fp does not follow frame pointer chains and always scans.

-callstack_conservative checks that each retaddr follows a call that
targets the prior callstack frame.

Needless to say, these add overhead, so they are not on by default,
although -no_callstack_use_fp's overhead is small enough that if we feel
comfortable with it we could turn it on by default.

SVN-Revision: 1642