Fixes issue 1409
Adds DRMF documentation.

Adds our 3 headers to what we pass to doxygen.

Fixes a number of errors in our doxygen comments that we never caught
before b/c we weren't actually building them into the docs!

Clarifies the types of some umbra parameters to simplify the docs and the API:
 - Clarified the final parameter to umbra_get_shadow_memory_type() and
   umbra_shadow_memory_is_shared() to be #umbra_shadow_memory_type_t
   rather than uint.
 - Clarified the 2nd parameter to umbra_create_shadow_memory() to be
   #umbra_shadow_memory_flags_t rather than uint.

Relies on DR changes to rename "Files" and "Data Structures" to have a
"DRMF" prefix, to avoid confusing docs which are mostly about the
Dr. Memory tool.

SVN-Revision: 1665