Fixes issue 1352
Adds buffering of writes to drstrace.  It still flushes after every
pre and post syscall, so the buffering just combines each argument.
That still makes a measureable difference in some workloads.

Relies on dr_write_file() being re-routed to the console.
Updates DR to r2523 to get that support.



SVN-Revision: 1700