i#1440 MacOS DrSyscall: initial BSD syscall table
Add a script mksystable_macos.pl and its initial output to fill in the
MacOS BSD syscall table.  All the memory parameters are marked as output
for this commit.  They will be manually updated in future changes.

I put the table in a separate file to avoid the enormous source files we
have for Windows and Linux.

SVN-Revision: 1728