i#1453 first pass: generalize i#471's fld;fstp handling to also handle a "load memory from address A into xmmN; store xmmN into address B". It uses the same scheme of marking B as bitlevel to come back to the slowpath for the actual shadow copy A->B.
Also generalizes it to handle arbitrary instructions in between the
load and store.  This is slightly risky as there could be aliased
memory there that would then cause the shadow xfer at the store to
copy the wrong shadow value, but we live with that for this stopgap
solution until we have full xmm shadowing.

SVN-Revision: 1745