Fixes issue 1453
Generalize the memory copy through fp/xmm even further by detecting
and handling the destination base register being set in between the
load and store.  This does have some risks and is not foolproof but it
does the job and we'll remove it once we have true shadowing of fp and
xmm registers (issue 471, issue 243).

This enables detection of the fp pattern used by gcc, so I enabled
that part of the app_suite float test.

Adds a xmm xfer test to registers.c.

SVN-Revision: 1746