Fixes issue 1453
This is also the first step for i#243:

Adds shadowing of xmm registers and propagation to and from memory, but
very limited intra-xmm operation propagation for now.

All xmm registers are shadowed, and we propagate to and from them on loads
and stores.  For now we only shadow on certain operations we know we can
handle.

The shadowing is indirected, with a pointer to the memory in one new field
in TLS.  Raises the spill slots to 6 for this pointer.  Adds support for
the indirection to the fastpath, using a 3rd scratch register for
simplicity, and re-loading every time.  This could be optimized in the
future.

Puts in place basic OR-combining for inter-xmm copies, but we need more
infrastructure to start mirroring xmm opcode data movements: that's for the
rest of i#243.

Generalizes the shadow register layer to handle 16-byte registers (and thus
4-byte shadow values).

Generalizes the slowpath to handle 16-byte registers.

Starts generalizing the fastpath to handle src size != dst size for
super-dword operations, but that is not yet complete.

Does not handle ymm registers yet.

Cleans up the incorrect use of mi->need_nonoffs_reg3 for cases that
actually need memref offsets by adding a separate field
mi->need_offs_early, allowing the xmm usage to request a 3rd scratch reg
without triggering unrelated mem xl8 interactions that assume it needs a
memory offset (we only support aligned in fastpath for now).

SVN-Revision: 1755