Fixes issue 1473
Adds shadowing and propagation for mmx registers, using the same shadow
indirection as for xmm registers.

Fixes check-definedness for xmm and mmx in the fastpath to properly load
indirected shadow values.

Does not yet propagate more than we already do for xmm, but i#243 covers
that as it will be the same work for mmx as for xmm.

Adds some tests to registers.c.

SVN-Revision: 1780