Fixes issue 1503
On Vista for a graphical app in cmd, have the frontend show the
summary, as there is no output from the client.

On Win8+ for a graphical app, we can't do the normal robust in-cmd
test, so if the SHELL env var is not set then we show the summary, as
there is no output from the client.  We end up with a double-summary
from cmd run within rxvt but we can live with that.

SVN-Revision: 1861