Fixes issue 1529
Adds a new option -check_uninit_blacklist which is a ,-separated list of
modules inside which we mark all dests as defined and do no uninit
checking.

I implemented this using callstack's per-module user_data to check at
module load.  For fastpath, stores in bb_info_t.  For slowpath we have to
look up every time, but I put in some caching, and I'm using callstack's
fast module lookup.  I had to use an extra lookup for the "medium path"
unfortunately.

Adds a feature of automatically adding any whole-module suppression to the
-check_uninit_blacklist, which is more efficient as it avoids the slowpath.

Adds two tests, one of the explicit option and one of the auto-add from a
suppression.

Adds a new section to the documentation explaining this new option.

SVN-Revision: 1898