Fixes issue 1547
Adds custom success and output handling for
NtQueryVirtualMemory.MemoryWorkingSetList.  This is too specific (depending
on both the query enum code and the value of the size param) to encode as a
general flag in a secondary table so it ended up as custom code.

Adds a test to app_suite.

SVN-Revision: 1927