Fixes issue 1568
Enables -delay_frees_stack by default, now that we have i#672 and
-free_max_frames is set to a low value (6 by default).

This is still a perf hit on cfrac, as well as on Chromium unit_tests, but
we're willing to live with that for the increased information in error
reports.

Adds documentation on tuning the performance of Dr. Memory.

SVN-Revision: 1967