Fixes issue 1593
Label use-after-free errors more prominently via a subtitle "of freed memory":

Error #1: UNADDRESSABLE ACCESS of freed memory: reading 1 byte(s)

This requires gathering and storing heap information up front and printing
it out separately when reporting an error.

We don't label an access to a freed redzone as a use-after-free, as
it can easily be an underflow from an adjacent live malloc.
We'll still list it as "N bytes beyond memory that was freed".
We do label access to freed padding.

Updates the docs and tests.

SVN-Revision: 1984