i#1009c#4: auto-register Dr. Memory as a Visual Studio External Tool
Adds a utility program that supports both registering and unregistering
Dr. Memory as a Visual Studio External Tool, for every version of Visual
Studio and Visual Studio Express detected on the machine (though VS2010
Express in Basic mode will not work).

Adds calls into the utility program for installing and uninstalling in the
NSIS package.

Updates the docs about the auto-install.

SVN-Revision: 2045