Transition to git: developer setup script and git hooks
Adds a developer setup script and git hooks to add aliases and
configurations for our workflow, including the use of submodules.  A
developer should run make/git/devsetup.sh once in each repository after
cloning.

The pre-commit hook includes a check for accidental rollback of the
DynamoRIO submodule, which can happen when pulling but not updating the
submodules (which the "pullall" alias helps to avoid).

Currently some of these scripts assume that unix utilities are on the
path.  We may need to augment them for better Windows support.