Add new code review front end: "git review"
Adds an alias "git review" that calls a new script git_review.sh.  It
provides simple updates of new patchsets to the same review, commit message
reviewing by uploading the precise message for each patchset, and
auto-adding of the review URL to the commit message.

The existing "git dcommit" now also calls git_review.sh to mark the review
as committed with its final diff and a pointer to the online commit, as
well as remove the current review issue marker.

This is directly based on the DynamoRIO script with minor modifications.
It would be nice to share the code but it is non-trivial unless we stop
supporting separately-built DynamoRIO.

Review-URL: https://codereview.appspot.com/187890043