i#1680 win8.1 crash: fix incorrect shadow dst write
Fixes the crash in i#1680 where our in-heap ignore-unaddr code incorrectly
writes directly to the shadow, which fails for indirected xmm destinations.
I was unable to add a non-hacky test.  If we get more info on why this
happened (I could not reproduce) we can augment the partial test I added
here in asmtest.c.

Fixes #1680

Review-URL: https://codereview.appspot.com/199660043