i#1699 malloc zones: add initial Mac malloc zone API support
Adds replacement functions for 7 of the 10 malloc zone API routines.
A malloc zone is essentially just a separate heap.  It has a layer of
indirection where the app can replace any of the allocator routines,
but we ignore that part as we want our own routines to always be in
place.

Adds a new test mac_zones.

Still TOOD: add support for aligned-alloc routines, but first we need
base support for that in our own allocator (i#94).

Review-URL: https://codereview.appspot.com/265860043