i#1722 crash: do not check unaddr for indirected shadow reg
For the unaddr ignore feature while inside wrapped heap routines, we
crash if we try to access an indirected shadow reg, due to an
un-encodable instruction.  As it does not make sense to check its
shadow value when it's not propagated, we instead lazily pull in the
next shadow, the mem shadow.

Fixes #1722

Review-URL: https://codereview.appspot.com/230530043