i#1726 port to ARM: port drsyscall to ARM, part 3
Adds ARM-specific syscalls to the Linux syscall table.  These include some
syscalls not present on x86 that are not strictly ARM-only, such as
SYS_pciconfig_{iobase,read,write} and SYS_send and SYS_recv.

Unfortunately I have no simple test so these are penciled in and not fully
tested for now.

Review-URL: https://codereview.appspot.com/244980043