i#1669 windows 10: update version checks and syscall numbers and tables
Updates version checks for Windows 10.

Adds new columns of Windows 10 x86, x64, and wow64 numbers for user32,
imm32, gdi32, kernel32, and ntdll to drsyscall_numx.h and adds selection of
the right set.

Adds a new column for Windows 10 to drsyscall_usercallx.h but it is not yet
updated.

Adds entries for new Windows 10 syscalls to the ntoskrnl, ntuser, and ntgdi
tables, with the proper number of arguments.  Filling in details of the
shape of those arguments is split off to i#1750.

Updates sysnums.pl which was used to generate the drsyscall_numx.h tables.

Updates the required DR version as we now need Windows 10 support from DR.

Review-URL: https://codereview.appspot.com/256640043