i#1153,i#1360,i#1750 new usercalls: fill in many new usercall names and numbers
Uses the new winsysnums -usercalls feature to fill in many new usercall
names, correcting the sequential numbers for others.  There are still some
unknown names and gaps but the numbers are much more complete now for win8,
win8.1, and win10.

Updates the corresponding tables with entries marked UNKNOWN as the return
type and parameter types are not fully known yet.

Updates the code to not add the secondary name by itself to the table when
it is UNKNOWN.

Review-URL: https://codereview.appspot.com/257540043