i#1753: fix NtUserMessageCall return value
Despite the ReactOS header in our wininc/ntuser.h, NtUserMessageCall does
not return a bool but rather returns the LRESULT for the message (instead
of that coming from an OUT param).

Also adjusts the success value for integer return types: if the return type
is not NTSTATUS we assume it succeeded.

Fixes #1753

Review-URL: https://codereview.appspot.com/260460043