i#1669 windows 10: drsyscall and drstrace test fixes
Relaxes a drsyscall test assert about not syscalls failing, as
NtQueryValueKey fails on Windows 10.

Adds better diagnostics for drstrace_unit_tests, which can fail
mysteriously with no message due to the hardcoded path when the test
suite is copied.

Changes drsyscall_test to query syscall 16, not 1 which is not present on
win10wow64 (xref i#1692).

Review-URL: https://codereview.appspot.com/260460044