i#1620 WIX integration: add DrMem to user PATH
The WIX installer now unconditionally (i#1761 covers adding checkboxes)
adds Dr. Memory to the user's PATH.  This *should* be auto-removed on
uninstall but that is not working: i#1762 covers fixing that.

This CL updates the docs to reflect the new automatic PATH editing.

The WIX installer now has enough of the features of the NSIS installer that
WIX support is declared complete.

Fixes #1620

Review-URL: https://codereview.appspot.com/265780043