i#94 aligned malloc: do not report memalign underflow as use-after-free
Adds a check for whether there's a free callstack, to avoid reporting a
memalign underflow accessing pre-alloc padding (which is artificially
freed) as a use-after-free error.  If !options.delay_frees_stack we label
it anyway: better to properly label a real use-after-free and have bad
labeling for a rare memalign underflow.

Review-URL: https://codereview.appspot.com/262250045