i#1734 Create Dr. Fuzz: Fix bugs in fuzzer and mutator
Fixes several minor bugs in the drmemory fuzzer, especially
state handling in multi-threaded targets, and determining
the current fuzzer state when the target is not reached on
the same thread that the target was registered on, or is
not registered at module init time. Also fixes bugs in
drfuzz_mutator, when mutating certain buffer sizes.

Review-URL: https://codereview.appspot.com/268810043