i#1734 Create Dr. Fuzz: Integrate drfuzz_mutator
Integrates the drfuzz_mutator into the Dr. Memory
fuzzer, adding command-line options for the
mutator descriptor, the random seed, and for
specifying a single input value (as a bug
checking utility). Adds a top-level doc entry
"Fuzz Mode" with full explanation of the new
options and the mutator's behavior. Adds new
tests for the various mutator modes.

Review-URL: https://codereview.appspot.com/265130043