i#1773: handle bus error during leak scan
Wraps the leak scan de-reference in a fast safe read to avoid a potential
bus error.  Performance tests on cfrac show that this has negligible
overhead.

Adds a test that triggers SIGBUS without this fix.  The test uses the
VALGRIND_DO_LEAK_CHECK annotation to avoid having to wait for process exit.

Fixes #1773

Review-URL: https://codereview.appspot.com/265220043