DRi#1818: use oauth2 authentication to the code review site
Updates DR to a86f827 for the new upload.py with token support.

Updates git_review to look for a token in $HOME/.codereview_dr_token and to
pass it along with --oauth2 to the new upload.py.

Updates the dcommit alias to upload prior to committing, and to upload
prior to removing the issue marker, to support aborting or failing on
authentication.

Review-URL: https://codereview.appspot.com/267630043