i#1822: RtlSetThreadPreferredUILanguages is a nosy heap routine too
Treats RtlSetThreadPreferredUILanguages just like we treated
RtlGetThreadPreferredUILanguages from i#1565: it is a "nosy" routine
that checks heap headers.  We thus make allocs inside it go native.

Improves on the i#1720 solution of handling a native alloc with a call
to RtlFreeHeap (instead of RtlpFreeHeap as in the original case) by
adding a hashtable that stores all native allocs, avoiding invalid
heap arg false positives.

Fixes #1822

Review-URL: https://codereview.appspot.com/280820043