i#111 64-bit: include 64-bit drmemory in release package
Adds official support for 64-bit Dr. Memory in the installation package,
with a message printed out that uninitialized reads are not yet detected.

Updates all of the README and documentation pages to state that 64-bit is
now (partially) supported, though we still recommend building applications
as 32-bit for uninitialized read detection.

Review-URL: https://codereview.appspot.com/282730043