i#1726 port to ARM: pattern mode by default in spill code
Changes the options for ARM to make pattern mode the default, with the same
message that we use for 64-bit.

Disables all of the custom register spilling code (none used for pattern
mode) in the refactored spill.c for ARM until we've ported that to use
drreg.

Updates the segment base access code to make it cross-platform by changing
get_own_seg_base() to directly call dr_get_dr_segment_base() instead of
storing that value in a TLS slot that is read via inline asm.

Review-URL: https://codereview.appspot.com/280470043