i#1838: add new option -malloc_callstacks
Adds a new option -malloc_callstacks to obtain malloc information for
mismatch errors when leaks are disabled (e.g., in light mode).
Unfortunately, the performance cost is enough that this option is off by
default.

Adds a new test that runs cs2bug with "-light -malloc_callstacks" and
ensures we get the "memory was allocated here" messages.

Adds a note to mismatch error reports that a re-run with -malloc_callstacks
will add the alloc callstack.

Fixes #1838

Review-URL: https://codereview.appspot.com/277560043