i#1844: add a code coverage feature to Dr. Memory
Adds a new option, -coverage, which leverages DRi#1829's drcovlib to
measure code coverage.  The results are output to a file in the regular
Dr. Memory log directory.

Adds a new page to the documentation explaining the code coverage option
and how to post-process the raw data.  Adds a copy of drcov2lcov to the
Dr. Memory release package.

Updates DR to a7b2511 for drcovlib.

Fixes #1844

Review-URL: https://codereview.appspot.com/286720043