i#1726 port to ARM: fix ARM-specific bugs in drfuzz repeats
We need to restore %lr in the post-fuzz callback to avoid drwrap reading
the wrong retaddr in the work it does prior to the pre-fuzz callback.  With
the wrong retaddr it typically ends up flushing part of the cache when it
didn't need to do anything.

Also updates DR to 52cb067 to fix a drwrap decorated-PC bug that can cause
crashes when repeating functions with drfuzz.

Review-URL: https://codereview.appspot.com/288850043