i#1879: handle asymmetric ARM sub-word ISA
Adds an explicit check for whether a 1-byte memref has a valid 2-byte form
(there are many instances where this is not the case for ARM and Thumb).
If not, falls back to using drutil_insert_get_mem_addr() followed by a
plain OP_ldrh.

Adds asmtest_arm as a new test to verify the fix.

Fixes #1879

Review-URL: https://codereview.appspot.com/296930043