i#1908: support reading syscall numbers from a text file
Tries to help future-proof Dr. Memory on Windows by adding support for
reading system call numbers from a text file that we can provide to
avoid requiring a full package release.  Also adds related features to
support attempts at continued execution.

First, adds spot-checks of several syscalls to identify whether the
built-in tables are correct.  When the built-in tables are not
correct, or we cannot identify the Windows version, attempts to read
the numbers from a text file in the bin directory.  If there is no
file, or the file's contents do not contain entries for the current
kernel, prints a message to the user to download a new file from
dynamorio.org.

Adds a new option -ignore_kernel which allows for continuing past the
message.  For this feature, we refactor the code to attempt to obtain
win32k.sys numbers from wrappers at module load time, including looking in
win32u.dll.  We also relax some asserts when running this way.

As the drsyscall library is the one reading the file, we document the
format and pass in a path to the file.

Changes the general order in which syscall numbers are obtained:
avoids bothering to read the wrappers when the tables are correct.

Review-URL: https://codereview.appspot.com/305350043