DRi#1991: only support OSX 10.9+
Due to inability to easily build binaries for 10.7, and the fact that Mac
in general needs less legacy support, we bump our minimum supported Mac
version to OSX 10.9 (Mavericks).

Updates DR to 759cb2b for i#1991: ensure flag changes apply to C++ targets,
to ensure that drsyms is built for the right OSX target.

Review-URL: https://codereview.appspot.com/306310043