i#1795 drreg: move xl8 slowpath clearing to inlined instru
Moves the clearing of translation address sharing when the slowpath is
entered (due to clobbering the register) from C code, where we need to
know the TLS slot (which is not drreg-friendly) and we need a 2-step return
path, to the inlined instrumentation at the return-from-slowpath point.
This increases the code cache size slightly but should enable removal of
the 2-step return and makes moving to drreg easier.

Review-URL: https://codereview.appspot.com/311520043