i#111 x64: 64-bit shadow slowpath
Generalizes register shadow operations on the slowpath for 64-bit.
Widens the eflags shadow to be 2 bytes to match the other registers, for
ease of propagation.

Clears the top 32 bits of a GPR when the bottom 32 bits are written and
adds an explicit test of this to asmtest.

Skips shadow memory on the leak scan and supports 2-byte shadow value
searches in umbra x64 to get leaks working for x64 shadow.

Fixes a -verbose 5 bug from changes in 33e1c26.

This gets 64-bit tests/free working with -no_fastpath -no_esp_fastpath.

Issue: 111