i#2041: use drutil_insert_get_mem_addr()
Switches to using drutil_insert_get_mem_addr() instead of DrM's custom impl
(which was the inspiration for drutil, but had gotten out of sync).  This
fixes bugs in handling rip-rel memrefs.

Relies on the latest DR f6e8e1c for changes to drutil_insert_get_mem_addr()
that guarantee when an extra scratch reg is used.

Fixes #2041