DRi#2764 Travis OSX: skip OSX for PR's
Skips running OSX for pull requests for now, since the Travis queue is
too long. We'll rely on the push to master to catch OSX issues and the
dev will have to revert or fix quickly at that point.

Re-writes the Travis build config from a matrix defined by os, compiler,
and env with exclusions to an explicit list of jobs, in order to use
conditional jobs. This is also clearer and easier to read.