i#2083 x64 reach: put DrM lib in low mem
As a short-term workaround for x64 reachability problems with shadow tables
that are currently directly accessed with displacements, we add to the
default DR options "-no_vm_base_near_app".  Long-term we need to get a
local scratch register and load the table base into it, which will be much
easier with drreg (#1795).

Tested on both Windows with x64 full mode enabled and on Linux in a distro
where this assert is hit.

Issue: #2083