Update non-Windows base version to 2.0
The version split when we added 64-bit full mode to UNIX and released
it as 2.0.x before it was ready for Windows.  The cronbuilds were
using the 1.11.x Windows number so we could share the same Github
version, but that just caused (justifiable) confusion.  Here we bump
the version to 2.0 for UNIX but leave it at 1.11 for Windows, which
will result in different git tags and cronbuilds until we get Windows
up to 2.0.