Allow package.cmake to specify preferred dbghelp and symsrv paths (#2173)
Adds two new CMake variables, DBGHELP_GLOB and SYMSRV_GLOB.  These are
added to the front of the search paths for dbghelp.dll and symsrv.dll.
This allows packaging to specify a glob for a prefered path for both
x86 and x64, which was not possible before.  This simplifies building
a package on a recent Windows version and having it work on an older
version.