i#2255: Build cronbuilds as official packages
Switches to a separate invocation of package.cmake for cronbuilds, to
make them look like release packages.  Cronbuilds on Travis and
Appveyor no longer run the same jobs as a regular build but instead
run one package.cmake job.

If the build number is 0, we omit it from the package name, rather
than appending it.

.travis.yml now takes in VERSION_NUMBER and uses it directly.  There
is no more TAG_SUFFIX: VERSION_NUMBER has to include the build number
if that's desired, and that will be part of the tag as well.  If
VERSION_NUMBER is not set it uses the default as before with a
cronbuild- prefix for the tag.

runsuite_wrapper.pl parses VERSION_NUMBER and if it has a -NNN it
passes that as the build= arg to package.cmake.  It also passes a
version= argument.

Fixes #2255